Remove PyEnchant, it won't be installed.
`README.rst` says "PyEnchant (optional)" but it's not. it will be installed no matter what.
Well it's in `setup.cfg`;

```
#!python

requires-dist =
    pyenchant
```

It cause some problems on platforms that doesn't allow applications to use python packages with `C` extension dependencies, Such as GAE, Heroku. So it's not possible to use `guess_language` over there.